requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.0.3
paddlex
fastapi
uvicorn
requests
pillow
numpy==1.26.4
opencv-python-headless==4.8.1.78